requirements.txt
marshmallow==3.7.1
marshmallow-jsonapi==0.23.2
alembic==1.4.0
amqp==2.5.2
aniso8601==6.0.0
bcrypt==3.1.6
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.7
Mako==1.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.1
requests==2.23.0
six==1.12.0
SQLAlchemy==1.3.13
urllib3==1.25.8
vine==1.3.0
Werkzeug==0.15.2
WTForms==2.2.1
zipp==3.0.0

